Fix deployment with disabled caching.
Currently the template only checks if dns_caching_dns is set, so if you set it to False it will still
add the root zone caching.
Meanwhile the named.root file will be only downloaded if dns_caching_dns evaluates to true.
This means that if you set dns_caching_dns to false it will *not* download the named.root file, but
will be added to the named.conf, making it so that 'named-checkconf' will fail as it tries to load
a non existing file.

Error:could not configure root hints from '/var/bind/named.root': file not found